bsp: simplify nrf53 specific implementation of radio bsp library
