+++ bsp/bsp.emProject
@@ -28,7 +28,7 @@
       project_dependencies="00bsp_timer_hf"
       project_directory="."
       project_type="Library" />
-    <file file_name="nrf/$(Lh2ImplementationFile)" />
+    <file file_name="nrf/lh2.c" />
     <file file_name="lh2.h" />
   </project>
   <project Name="00bsp_dotbot_rpm">
@@ -106,7 +106,7 @@
       project_dependencies="00bsp_clock"
       project_directory="."
       project_type="Library" />
-    <file file_name="nrf/$(RadioImplementationFile)" />
+    <file file_name="nrf/radio.c" />
     <file file_name="radio.h" />
   </project>
   <project Name="00bsp_rng">
