+++ otap/nrf5340dk-app-bootloader.emProject
@@ -42,7 +42,7 @@
     linker_memory_map_file="$(ProjectDir)/nRF5340_xxAA_Application_MemoryMap.xml"
     linker_output_format="hex"
     linker_section_placement_file="$(ProjectDir)/../../nRF/Setup/nRF5340_xxAA_Application_flash_placement.xml"
-    macros="BuildTarget=nrf5340dk-app;Lh2ImplementationFile=lh2.c;PwmImplementationFile=pwm.c;RadioImplementationFile=radio_nrf5340_app.c;DeviceFamily=nRF;Target=nRF5340_xxAA_Application"
+    macros="BuildTarget=nrf5340dk-app;PwmImplementationFile=pwm.c;DeviceFamily=nRF;Target=nRF5340_xxAA_Application"
     project_type="Executable"
     target_reset_script="Reset();"
     target_trace_initialize_script="EnableTrace(&quot;$(TraceInterfaceType)&quot;)" />
